pyproject.toml
[build-system]
requires = ["setuptools>=65.5", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tripa"
version = "0.2.0"
description = "Utilities to create, open, extend, and animate .tripa files."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
  { name = "Cognet Benjamin" },
]
dependencies = [
  "Pillow>=9.0",
]
keywords = ["tripa", "image", "art", "python"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Graphics",
]

[project.urls]
Homepage = "https://github.com/Hatsunefanboy/Tripatouilleur"
Repository = "https://github.com/Hatsunefanboy/Tripatouilleur"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tripa = [
  "tripa/forwards_tripa/images/*.py",
  "tripa/saved_tripa/*.tripa",
]
